chore: update controller names, namespaces and added create
